Fixed offset bugs in union.

+++ include/set/ops/union.hpp
@@ -7,18 +7,18 @@ namespace ops{
       const uint32_t *a_index = (uint32_t*) A_in->data;
       const uint32_t *b_index = (uint32_t*) B_in->data;
       
-      const uint64_t * const A = (uint64_t*)(A_in->data+sizeof(uint32_t));
-      const uint64_t * const B = (uint64_t*)(B_in->data+sizeof(uint32_t));
-      const size_t s_a = ((A_in->number_of_bytes-sizeof(uint32_t))/sizeof(uint64_t));
-      const size_t s_b = ((B_in->number_of_bytes-sizeof(uint32_t))/sizeof(uint64_t));
+      const uint64_t * const A = (uint64_t*)(A_in->data+sizeof(uint64_t));
+      const uint64_t * const B = (uint64_t*)(B_in->data+sizeof(uint64_t));
+      const size_t s_a = ((A_in->number_of_bytes-sizeof(uint64_t))/sizeof(uint64_t));
+      const size_t s_b = ((B_in->number_of_bytes-sizeof(uint64_t))/sizeof(uint64_t));
 
       const bool a_big = a_index[0] > b_index[0];
-      const uint32_t start_index = (a_big) ? a_index[0] : b_index[0];
-      const uint32_t a_start_index = (a_big) ? 0:(b_index[0]-a_index[0]);
-      const uint32_t b_start_index = (a_big) ? (a_index[0]-b_index[0]):0;
+      const uint64_t start_index = (a_big) ? a_index[0] : b_index[0];
+      const uint64_t a_start_index = (a_big) ? 0:(b_index[0]-a_index[0]);
+      const uint64_t b_start_index = (a_big) ? (a_index[0]-b_index[0]):0;
 
-      const uint32_t end_index = ((a_index[0]+s_a) > (b_index[0]+s_b)) ? (b_index[0]+s_b):(a_index[0]+s_a);
-      const uint32_t total_size = (start_index > end_index) ? 0:(end_index-start_index);
+      const uint64_t end_index = ((a_index[0]+s_a) > (b_index[0]+s_b)) ? (b_index[0]+s_b):(a_index[0]+s_a);
+      const uint64_t total_size = (start_index > end_index) ? 0:(end_index-start_index);
 
       //16 uint16_ts
       //8 ints
@@ -52,7 +52,7 @@ namespace ops{
     }
   }
   inline void set_union(Set<bitset> *A_in,Set<pshort> *B_in){
-    uint64_t* A = (uint64_t*)(A_in->data+sizeof(uint32_t));
+    uint64_t* A = (uint64_t*)(A_in->data+sizeof(uint64_t));
     const uint32_t * const s_index_p = (uint32_t*)A_in->data;
     const uint32_t start_index = (A_in->number_of_bytes > 0) ? s_index_p[0]:0;
 
@@ -88,7 +88,7 @@ namespace ops{
     set_union(B_in,A_in);
   }
   inline void set_union(Set<bitset> *A_in,Set<variant> *B_in){
-    uint64_t* A = (uint64_t*)(A_in->data+sizeof(uint32_t));
+    uint64_t* A = (uint64_t*)(A_in->data+sizeof(uint64_t));
     const uint32_t * const s_index_p = (uint32_t*)A_in->data;
     const uint32_t start_index = (A_in->number_of_bytes > 0) ? s_index_p[0]:0;
 
@@ -104,7 +104,7 @@ namespace ops{
     set_union(B_in,A_in);
   }
   inline void set_union(Set<bitset> *A_in,Set<bitpacked> *B_in){
-    uint64_t* A = (uint64_t*)(A_in->data+sizeof(uint32_t));
+    uint64_t* A = (uint64_t*)(A_in->data+sizeof(uint64_t));
     const uint32_t * const s_index_p = (uint32_t*)A_in->data;
     const uint32_t start_index = (A_in->number_of_bytes > 0) ? s_index_p[0]:0;
